Extracts `hideWhenChildBlocks` component for use on `InnerBlocks`
