+++ packages/block-editor/src/components/inner-blocks/hide-when-child-blocks.js
@@ -1,0 +1,32 @@
+/**
+ * WordPress dependencies
+ */
+import { compose, ifCondition } from '@wordpress/compose';
+import { withSelect } from '@wordpress/data';
+
+/**
+ * Internal dependencies
+ */
+import withClientId from './utils/with-client-id';
+
+export const hideWhenChildBlocks = function( props ) {
+	return props.children;
+};
+
+export default compose( [
+	withClientId,
+	withSelect( ( select, { clientId } ) => {
+		const {
+			getBlock,
+		} = select( 'core/block-editor' );
+
+		const block = getBlock( clientId );
+
+		return {
+			hasChildBlocks: !! ( block && block.innerBlocks.length ),
+		};
+	} ),
+	ifCondition( ( { hasChildBlocks } ) => {
+		return ! hasChildBlocks;
+	} ),
+] )( hideWhenChildBlocks );
